+++ R/utils.R
@@ -66,3 +66,14 @@ copy_create_dir <- function(from, to, ...) {
     file.copy(from = from[i],  to = to[i], ...)
   }
 }
+
+scoped_tempdir <- function (code){
+
+  new <- tempfile(pattern = "file", tmpdir = tempdir(), fileext = "")
+  dir.create(new)
+  on.exit(unlink(new, recursive = TRUE), add = TRUE)
+  old <- setwd(dir = new)
+  on.exit(setwd(old))
+  force(code)
+}
+