+++ NEWS.md
@@ -1,3 +1,18 @@
+# theorytools 0.1.3
+
+* Fix issue raised by Prof. Brian Ripley:
+  download_theory() examples fail if no internet.
+* Add support for Python functions
+* Add download_huggingface() to download models from huggingface
+* Add get_embeddings() to get vector embeddings using huggingface model
+* Add code() to code qualitative text entries with support of cosine similarity
+  or string distance
+* Add add_level() for recursive coding of qualitative text entries
+* Add similarity_stringdist() to calculate similarity between character vectors
+  based on string distance
+* Add similarity_cosine() to calculate cosine similarity between vector
+  embeddings, see get_embeddings()
+
 # theorytools 0.1.2
 
 * Minor bugfixes